pyproject.toml
[tool.poetry]
name = "skypro-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
ipython = "^8.27.0"
django-stubs = "^5.0.4"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.1"
drf-spectacular = "^0.27.2"
stripe = "^10.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
